refactor: update import paths for country data and image assets across multiple components
