+++ src/front/modules/miscellaneous/4geeks-defatult-files/Single.jsx
@@ -1,7 +1,7 @@
 // Import necessary hooks and components from react-router-dom and other libraries.
 import { Link, useParams } from "react-router-dom";  // To use link for navigation and useParams to get URL parameters
 import PropTypes from "prop-types";  // To define prop types for this component
-import rigoImageUrl from "../shared/assets/img/rigo-baby.jpg"  // Import an image asset
+import rigoImageUrl from "../../../shared/assets/img/rigo-baby.jpg"  // Import an image asset
 import useGlobalReducer from "../../../shared/hooks/useGlobalReducer";  // Import a custom hook for accessing the global state
 
 // Define and export the Single component which displays individual item details.
